Split jsonparser functionality, added filtering based on map/flight path
